Remove unused code for GetPreferences hook callback
Removed unused code for GetPreferences hook callback; at
the time I was considering writing a visual editor interface
and including a user preference for switching between editing direct
source and a visual editor. However, the idea of a visual editor
right now is not the main focus (and we should probably do away
from even more user preferences than there already is).

Note: The hook callback for GetReservedUsernames is eventually planned
to be used for migrating wikitext-based navigation templates to
structured navigations, so we're keeping that here at least.