Hygiene: Separate structure + content by giving CSS class to <dl> element
This allows the styles to be independent of the HTML structure;
as in, the styles do not have to know what kind of HTML element
is represented by the class, it just knows that it must give
these styles to the specified class.

This also allows for more readable code.