Hygiene: Narrow scope of constants
This narrows the scopes of constants; only things that truly
need access to them have access. Where needed, this is done
by marking these constants as private (feature of assigning
visibility to constants was introduced in PHP 7.1.0 [1]).

Constants for service keys have been moved into their own file,
instead of under the Services class so there wouldn't be an
interdependency between the service wiring file and Services class.

The parts that are only absolutely for dealing with *just* services
are now under a new sub-namespace, \StructuredNavigation\Services,
to help with organizing.

This (basically) reverts b78c6c9, for reasons outlined above and
in #22. This fixes #22.

References:
 - [1]:
   - http://php.net/manual/en/language.oop5.visibility.php#language.oop5.visiblity-constants
   - http://php.net/manual/en/language.oop5.constants.php (example #5)
   - RFC: https://wiki.php.net/rfc/class_const_visibility