Documentation: Improve README, new guide on using with TemplateStyles
It's now officially recommended that if you'd like to consider styling
your structured navigation, that you install TemplateStyles on the side.
If you want to learn how to use it with TemplateStyles, please read the
guide at docs/Styling-with-TemplateStyles.md.

NOTE: StructuredNavigation integrates nicely with TemplateStyles but does
**not** have a hard software dependency on it.

Other things done:
 - Added a contributing section to README.md
 - Added a license section to README.md