Hygiene: Remove need for 'name' property in navigation JSON
The 'name' property in a navigation was in the original implementation
for a navigation, but after some thinking this property is redundant.
Access the title label instead.

This change IS backwards-compatible on your wiki, but you should remove
the 'name' property from your existing navigations since it's no longer needed.