New empty state screen for when visiting a non-existing navigation
This hooks on to the BeforeDisplayNoArticleText hook to create a custom
empty state screen. This is designed to provide more engagement to the
end-user with a more prominent CTA to guide them to creating a new
navigation page.

Fixes #45. Screenshot is included in the task of the new empty state
screen.

Suggestions, feedback, and comments are welcome! You can provide any
by submitting a new issue.