Hygiene: More cleanup by removing some unused code
These were surfaced into the initial commit, as I was planning to use
these at some point for the extension. However, this hasn't quite happened
yet for a while, so it's better to remove them for now. If need be, these
can be always be added back again in the future.