License: Relicense everything under MIT
After reconsidering, I ultimately believe the MIT is a better
license as it's ultimately more permissible.

Also went ahead and relicensed the library code to be under the
MIT license too - it didn't make sense for it to be under the
GPL-2.0-or-later license when it's /meant/ to be reusable!
My apologies. :(