Hygiene: Make JsonEntityFactory use NavigationTitleValue + other cleanup
JsonEntityFactory now takes a NavigationTitleValue for the
constructor, which was better than before because classes
had to inject both at the same time,

The responsibility of title resolving has now also been placed
to the JsonEntityFactory. The factory will attempt to create
a JsonEntity object, but will fail early (return false) if the
given title doesn't exist. This is the right place for this
responsibility to be (or at least probably).

Fixes #56.