Add $wgStructuredNavigationEnableCopyEmbedFeature
Adds a configuration setting for deciding whether or not to enable this feature. By default it is false. You can enable it by setting this to true, but this feature will **only** work if you're running MW-1.34 alpha or above.

This configuration setting is temporary; it will by removed and the feature will be enabled by default when MW-1.34 officially releases.