Enhance UX when a user visits the "nav not found" screen
If the user visits a navigation page that doesn't exist *and* is not
logged in, they will be prompted to log in/create an account to create
the navigation. The button URL will also prompt them to Special:UserLogin,
which will redirect them back to the page after they have finished logging in.

The title message key has also been modified to include the name of the
navigation when the user is shown this screen.

Change-Id: I4735b0d2ce55f9efb442aa0dc4a1a22016f73e9f