Decouple from User and use the UserIdentity interface instead
User is a large object, and quite too large for our needs. Instead,
only require its lightweight interface, MediaWiki\User\UserIdentity.

This interface was introduced in 1.31.

The isRegistered() method, (which is an alias for isLoggedIn(), and
per the PHP doc block, is a more accurate description of what we're
actually checking) is introduced in the UserIdentity interface in 1.34.

Change-Id: I8bf9615ed567d57c0f00c77342975e338987349d