Make Action and REST API modules disabled by default, are experimental
These API modules are both very experimental. To make this clearer,
a configuration setting is added called:
$wgStructuredNavigationEnableExperimentalAPI.

By default, it is set set to false, which by default will make
these API modules not accessible to users. These are currently
still in heavy development, and should be acknowledged by the
site administrator or developer that enables this feature
when they set this to true.

Note: This also adds quite a bit of maintenance too, so
hopefully we can have this removed ASAP in the future too...
This is not meant to be kept for very long, but this
also depends on long how it will take to make sure these
modules are stable.

Change-Id: I854792ce749cf7bc6a76ac5854947ec3399f2758